Abs fix and gitignore change

+++ SHPullAcrossViewControllerExample/SHPullAcrossViewControllerExample/SHPullAcrossViewController.m
@@ -267,7 +267,7 @@
     CGFloat velocity = [recognizer velocityInView:self.pullAcrossView].x;
     NSTimeInterval duration;
     SHPullAcrossVCPosition position;
-    if(fabsf(velocity) > self.minimumPanVelocity)
+    if(fabs(velocity) > self.minimumPanVelocity)
     {
         if(velocity < 0)
         {
@@ -277,7 +277,7 @@
         {
             position = SHPullAcrossVCPositionClosed;
         }
-        duration = fabsf([self _distanceRemainingToPosition:position] / velocity);
+        duration = fabs([self _distanceRemainingToPosition:position] / velocity);
     }
     else
     {
